refactor: extract project/section/task event handling into delegates
